+++ package.json
@@ -8,16 +8,6 @@
   "main": "lib/index.js",
   "typings": "lib/index.d.ts",
   "module": "lib/es6/index.js",
-  "scripts": {
-    "check": "tsc -p .",
-    "lint": "tslint -p . --verbose",
-    "mocha": "mocha -r ts-node/register test/*.ts",
-    "test": "npm run lint && npm run mocha && npm run docs",
-    "prebuild": "rm -rf lib",
-    "build": "tsc -p ./tsconfig.build.json && tsc -p ./tsconfig.build-es6.json",
-    "fix-prettier": "prettier --write \"{src,test}/**/*.ts\"",
-    "docs": "docs-ts"
-  },
   "repository": {
     "type": "git",
     "url": "https://github.com/gcanti/elm-ts.git"
@@ -28,6 +18,28 @@
     "url": "https://github.com/gcanti/elm-ts/issues"
   },
   "homepage": "https://github.com/gcanti/elm-ts",
+  "tags": [
+    "typescript",
+    "elm",
+    "fp-ts"
+  ],
+  "keywords": [
+    "typescript",
+    "elm",
+    "fp-ts"
+  ],
+  "scripts": {
+    "check": "tsc -p .",
+    "lint": "tslint -p . -t verbose",
+    "jest": "jest --ci --coverage",
+    "pretest": "npm run check && npm run lint",
+    "test": "npm run jest",
+    "posttest": "npm run docs",
+    "prebuild": "rm -rf lib",
+    "build": "tsc -p ./tsconfig.build.json && tsc -p ./tsconfig.build-es6.json",
+    "fix-prettier": "prettier --write \"{src,test}/**/*.ts\"",
+    "docs": "docs-ts"
+  },
   "dependencies": {
     "axios": "^0.17.1",
     "history": "^4.7.2"
@@ -42,7 +54,7 @@
   "devDependencies": {
     "@types/blessed": "0.1.8",
     "@types/history": "4.6.2",
-    "@types/mocha": "^5.2.5",
+    "@types/jest": "24.0.18",
     "@types/node": "7.0.8",
     "@types/react": "16.0.27",
     "@types/react-dom": "16.0.3",
@@ -51,26 +63,38 @@
     "docs-ts": "^0.2.0",
     "fp-ts": "^2.0.2",
     "io-ts": "^2.0.0",
-    "mocha": "^5.2.0",
+    "jest": "24.9.0",
     "monocle-ts": "^0.5.2",
     "prettier": "^1.18.2",
     "react": "^16.8.6",
     "react-blessed": "^0.5.0",
     "rxjs": "^6.5.2",
+    "ts-jest": "24.1.0",
     "ts-node": "3.3.0",
     "tslint": "^5.18.0",
     "tslint-config-standard": "^8.0.1",
     "typescript": "^3.5.3",
     "webpack": "^4.26.0"
   },
-  "tags": [
-    "typescript",
-    "elm",
-    "fp-ts"
-  ],
-  "keywords": [
-    "typescript",
-    "elm",
-    "fp-ts"
-  ]
+  "jest": {
+    "preset": "ts-jest",
+    "globals": {
+      "ts-jest": {
+        "diagnostics": true
+      }
+    },
+    "bail": true,
+    "coveragePathIgnorePatterns": [
+      "<rootDir>/test/",
+      "<rootDir>/node_modules/"
+    ],
+    "coverageReporters": [
+      "text"
+    ],
+    "roots": [
+      "<rootDir>/test/"
+    ],
+    "testMatch": null,
+    "testRegex": "(\\.|/)(test|spec)\\.tsx?$"
+  }
 }
